fixed division, added zoom

+++ sem2/Homework5/task2/src/main/java/group144/stepyrev/Controller.java
@@ -69,7 +69,7 @@ public class Controller {
                 result = first * second;
                 break;
             case "/":
-                result = first / second;
+                result = (double)first / (double)second;
                 break;
         }
